+++ Android/App_New/App_Mix_Manual/src/main/java/com/zktony/android/data/entities/Program.kt
@@ -29,12 +29,12 @@ data class Program(
     /**
      * 开始浓度
      */
-    var startRange: Double = 0.0,
+    var startRange: Int = 0,
 
     /**
      * 结束浓度
      */
-    var endRange: Double = 0.0,
+    var endRange: Int = 0,
 
     /**
      * 常用厚度
@@ -45,7 +45,7 @@ data class Program(
     /**
      * 促凝剂体积
      */
-    var coagulant: Double = 0.0,
+    var coagulant: Int = 0,
 
     /**
      * 胶液体积
